Let django manage the SPA resources
Handling static resources nicely is difficult without including some
kind of build stage, but Django's collectstatic is basically a build
stage so use that as a processor. The main change here is that the SPA
index is served as a django template so sttaic paths can get subsituted
in, and a JS function for making partial paths point to the right place.